Implemented 2 APIs and Modified 1 API   1) Implemented POST /api/fm/outlets/createOutlet with State, City & Area ID-Based Address Validation, Merchant-Specific Bank Account Validation, Outlet Registration, Address Creation, Bank Details, Operating Days Configuration, and Outlet User Creation.  2) Implemented PUT /api/fm/outlets/updateOutletDetailsByMerchant with Complete Outlet Management Support Including Outlet Information, Contact Details, Address, Bank Details, Location, Operating Days Updates, and Merchant-Specific Bank Account Validation.  3) Modified GET /api/fm/outlets/getOutletDetails to Return Enhanced Outlet Information Including Outlet Email, Alternate Outlet Phone, Bank Details, Complete Address Information, Outlet Timings, Categories, Products, and Favourite Status Based on User Type.

+++ foodandmart/src/main/java/com/jippy/foodandmart/controller/FmMerchantController.java
@@ -42,6 +42,7 @@ public class FmMerchantController {
         return ResponseEntity.ok(FmApiResponse.success("Merchants fetched", merchantService.getAllMerchants()));
     }
 
+//    ------------------------------------------------------------
     /** GET /api/merchants/{id} */
     @GetMapping("/{id}")
     public ResponseEntity<FmApiResponse<FmMerchantDto>> getMerchantById(@PathVariable Integer id) {
@@ -78,8 +79,7 @@ public class FmMerchantController {
     }
 
     // update merchant with bank details
-    @Operation(
-            summary = "Update Merchant Profile",
+    @Operation(summary = "Update Merchant Profile",
             description = "Update merchant and bank details"
     )
     @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "Merchant updated successfully")
@@ -95,16 +95,17 @@ public class FmMerchantController {
 
         return ResponseEntity.ok(updated);
     }
+
+//    --------------------------------------------------------------------------
     //    get merchant details with bank details
-    @Operation(
-            summary = "Get Merchant Profile with Bank Details",
+    @Operation(summary = "Get Merchant Profile with Bank Details",
             description = "Fetch merchant details along with bank information using merchant ID"
     )
     @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "Merchant fetched successfully")
     @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "404", description = "Merchant not found")
     @GetMapping("/getMerchantProfile")
     public ResponseEntity<FmMerchantWithBankDto> getMerchantProfile(
-            @RequestParam Long merchantId) {
+            @RequestParam Integer merchantId) {
         log.info("Fetching merchant profile for merchantId: {}", merchantId);
         FmMerchantWithBankDto response =
                 merchantService.getMerchantWithBank(merchantId);
